Feature: exercises can be added to workouts

+++ src/main/java/com/dimitarrradev/workoutScheduler/exercise/dto/TrainingSetsServiceModel.java
@@ -3,7 +3,7 @@ package com.dimitarrradev.workoutScheduler.exercise.dto;
 public record TrainingSetsServiceModel(
         Long id,
         String exerciseName,
-        Integer sets,
+        Long sets,
         Integer minReps,
         Integer maxReps,
         Double weight,
